fix(sd34): AT-34-E6-001 wave 25 gate lane C -- clippy root 86->0, desktop 12->0, ceilings 50/7->0/0
Re-measured before touching anything, per the dispatch brief's own instruction. Root's fresh
measurement reproduced the review's "86" exactly, but cargo clippy --locked --tests actually
exited 101: a hard clippy::invisible_characters compile error (three literal U+00AD soft-hyphen
bytes in monster_codex/spell_list.rs, landed in SD-32, predates this bundle) blocked lib/lib-test
compilation, which means every tests/*.rs integration target and src/bin/*.rs binary's own test
build was never reached by this stage in this bundle's history -- the fable-review "86" and the
wave-23 gate-sweep's re-confirmed "86" were both, unknowingly, lib+lib-test-only counts. Fixed the
three bytes to regular hyphens first, which is what let clippy reach the true, full population for
the first time.

Root: 86 -> 0. cargo clippy --fix closed 60 mechanically; the rest hand-fixed: 2 genuinely-dead
functions deleted (confirmed superseded via git log -S and crate-wide grep, not assumed) with 3
now-redundant dedicated tests; a 7-file copy-pasted dead test helper deleted from all 7; 2 module
constants confirmed test-fixture-only and gated #[cfg(test)]; 6 type aliases for type_complexity;
6 narrow, individually-commented #[allow(clippy::too_many_arguments)] (no blanket/crate-level
allow -- every flagged function's params are real domain inputs, a signature refactor across an
~80k-line compute file is out of this cycle's scope).

Desktop: measured 12 (already down from the review's stale 25 via wave 24's own landed fixes) ->
0. cargo clippy --fix closed 3; 5 clone-to-std::slice::from_ref; 3 large_enum_variant warnings
closed by boxing both corpus_derived and (newly) snapshot across all three response enums --
boxing corpus_derived alone only closed 504->248 bytes, still over threshold; every
construction/destructure site updated compiler-error by compiler-error; 1 narrow, commented
#[allow(dead_code)] (confirmed an original SD-23 design element via git log -S, not a bloat-removal
orphan).

Both ceilings tightened, not raised: BASELINE_CLIPPY_WARNINGS_ROOT/DESKTOP 50/7 -> 0/0
(scripts/verify-baselines.env). scripts/verify.sh --only clippy -> PASS clippy (root:0 desktop:0
warnings, 0 errors), re-confirmed twice.

cargo test --locked --no-run: workspace exit 0 (589 binaries), apps/desktop/src-tauri exit 0.
BASELINE_CORPUS_LITERAL_RECORDS 26500->48708 (the brief's stretch goal) confirmed already landed
by the wave-23 gate-remediation closing sweep -- no action needed this cycle.

Receipt: docs/release/SD-34-book-completion/artifacts/epic-6-closure/AT-34-E6-001_gate-lane-c_wave25_cycle_receipt.md

+++ src/rules_core/rules_tables/crb/weapon_tables.rs
@@ -362,7 +362,7 @@ mod tests {
         assert_eq!(longsword.critical_threat_range_width, 2, "raw corpus CRITRANGE");
         assert_eq!(weapon_critical_threat_low(longsword), 19, "21 - 2");
         assert_ne!(
-            u8::from(longsword.critical_threat_range_width),
+            longsword.critical_threat_range_width,
             weapon_critical_threat_low(longsword),
             "width and low bound must not be conflated"
         );
